+++ Firmware/XPB-PIO/lib/lvgl_ui/lvgl_ui.h
@@ -6,7 +6,7 @@
 #include "FreeRTOS.h"
 #include "task.h"
 #include "xpb_color_palette.h"
-#include "lvgl_widget.h"
+#include "lvgl_ui_widget.h"
 
 lv_obj_t* dataMonitor_initUI(int rotation);
 lv_obj_t* dataMonitorCount_initUI(int rotation, uint8_t channel);
